+++ NEWS.md
@@ -30,6 +30,8 @@ rsat 0.1.15 (2021-08-18)
   * Function name change `cloud_mask()` -> `rsat_cloudMask()`
   * Function name change `smoothing_images()` -> `rsat_smoothing_images()`
   * Function name change `list_data()` -> `rsat_list_data()`
+  * Function name change `get_raster()` -> `rsat_get_raster()`
+  * Function name change `get_stars()` -> `rsat_get_stars()`
 
 ### DOCUMENTATION